Use polling for submit button
Polling seems to work better than using the method of adding and removing edge detections, as seen in the 'danilo' branch.
This method still times out at 2s, so if the user presses for more than 2s, it will have no effect.
The program will stop analysing the click after 2s or when the user clicks, whichever happens sooner.

This can be inefficient, but since the Pi does not have to be doing anyhting while the submit button is being pressed, it is ok.

This method is better than wating for the user to release the button to know what kind of click it is, as the user does not know how long to press for.